backend/pystagram/paginators.py
from rest_framework.pagination import PageNumberPagination


class SmallPagination(PageNumberPagination):
    page_size = 20
    max_page_size = 60
    page_size_query_param = 'page_size'


class LargePagination(PageNumberPagination):
    page_size = 50
    max_page_size = 100
    page_size_query_param = 'page_size'
