backend/requirements.txt
Django==4.1.2
Pillow==9.2.0
argon2-cffi==21.3.0
python-decouple==3.6
django-cleanup==6.0.0
djangorestframework==3.14.0
django-cors-headers==3.13.0
